fix: update handler name
