+++ src/Adapter/Shipment/QueryHandler/GetShipmentProductsHandler.php
@@ -26,17 +26,17 @@
 
 namespace PrestaShop\PrestaShop\Adapter\Shipment\QueryHandler;
 
-use PrestaShop\Decimal\DecimalNumber;
+use OrderDetail;
 use PrestaShop\PrestaShop\Core\CommandBus\Attributes\AsQueryHandler;
 use PrestaShop\PrestaShop\Core\Domain\Shipment\Exception\ShipmentNotFoundException;
-use PrestaShop\PrestaShop\Core\Domain\Shipment\Query\GetOrderShipments;
-use PrestaShop\PrestaShop\Core\Domain\Shipment\QueryHandler\GetOrderShipmentForViewingHandlerInterface;
-use PrestaShop\PrestaShop\Core\Domain\Shipment\QueryResult\OrderShipment;
+use PrestaShop\PrestaShop\Core\Domain\Shipment\Query\GetShipmentProducts;
+use PrestaShop\PrestaShop\Core\Domain\Shipment\QueryHandler\GetShipmentProductsHandlerInterface;
+use PrestaShop\PrestaShop\Core\Domain\Shipment\QueryResult\OrderShipmentProduct;
 use PrestaShopBundle\Entity\Repository\ShipmentRepository;
 use Throwable;
 
 #[AsQueryHandler]
-class GetOrderShipmentsForViewingHandler implements GetOrderShipmentForViewingHandlerInterface
+class GetShipmentProductsHandler implements GetShipmentProductsHandlerInterface
 {
     public function __construct(
         private readonly ShipmentRepository $repository,
@@ -44,37 +44,31 @@ class GetOrderShipmentsForViewingHandler implements GetOrderShipmentForViewingHa
     }
 
     /**
-     * @param GetOrderShipments $query
+     * @param GetShipmentProducts $query
      *
-     * @return OrderShipment[]
+     * @return OrderShipmentProduct[]
      */
-    public function handle(GetOrderShipments $query)
+    public function handle(GetShipmentProducts $query)
     {
-        $shipments = [];
-        $orderId = $query->getOrderId()->getValue();
+        $shipmentProducts = [];
+        $shipmentId = $query->getShipmentId()->getValue();
 
         try {
-            $result = $this->repository->findByOrderId($orderId);
+            $result = $this->repository->findOneBy(['id' => $shipmentId]);
         } catch (Throwable $e) {
-            throw new ShipmentNotFoundException(sprintf('Could not find shipment for order with id "%s"', $orderId), 0, $e);
+            throw new ShipmentNotFoundException(sprintf('Could not find shipment with id "%s"', $shipmentId), 0, $e);
+        }
+        if (!empty($result)) {
+            foreach ($result->getProducts() as $product) {
+                $orderDetail = new OrderDetail($product->getOrderDetailId());
+                $shipmentProducts[] = new OrderShipmentProduct(
+                    $product->getOrderDetailId(),
+                    $product->getQuantity(),
+                    $orderDetail->product_name,
+                );
+            }
         }
 
-        foreach ($result as $shipment) {
-            $shipments[] = new OrderShipment(
-                $shipment->getId(),
-                $shipment->getOrderId(),
-                $shipment->getCarrierId(),
-                $shipment->getAddressId(),
-                new DecimalNumber((string) $shipment->getShippingCostTaxExcluded()),
-                new DecimalNumber((string) $shipment->getShippingCostTaxIncluded()),
-                $shipment->getProducts()->count(),
-                $shipment->getTrackingNumber(),
-                $shipment->getShippedAt(),
-                $shipment->getDeliveredAt(),
-                $shipment->getCancelledAt(),
-            );
-        }
-
-        return $shipments;
+        return $shipmentProducts;
     }
 }
